Added the feature of tagging nodes with 'tested' attribute.
The 'tested' attribute will identify if a node was traversed by a test suite that was used when generating the dynamic call graph using gprof. All nodes from the cflow call graph have the 'tested' attribute set to False and all nodes from the gprof call graph thathave either a caller or a callee have the 'tested' attribute set to True. When the two call graphs are merged in the value of 'tested' is effectively set to True if a cflow node, that also appears in the gprof call graph, was tested.