Added the feature of including edge weights.
Edges now have a weight associated with them. An edge from the cflow call graph gets a weight of 1 and that from a gprof graph gets a 0.5. In other words, an edge that is tested is given a lower weight than that has not been tested.
Methods that compute the page rank of the nodes now take the weight into consideration. A node has a lower page rank if the weight of the in-link edge is low. The assumption her is that a node is less 'popular' to an attacker if it has been tested and thus the page rank of such a node is lower.