Bug in identifying the name field in gprof output
The field containing the name of the function and
the file containing the function in a line from
the gprof call graph begins at column 46 for line
that represents the function in question. However,
for its callers and callees the field begins at
column 50. Change adds code to differentiate the
types of lines when substringing.

Change fixed #46