Introduced the notion of granularity
With granularity, a node on the call graph can behave differently. For
instance, a node with granularity of function will have both the
function name and the name of the file containing the function as its
identity and a node with granularity of file will have only the name of
the file as its identity. The identity of the node controls the way in
which networkx adds nodes to a graph.